Fixed several warnings reported by VC6. Fixed songs not really supporting more than 64k...

+++ fmopl.cpp
@@ -1693,7 +1693,7 @@ static void OPLWriteReg(FM_OPL *OPL, int r, int v)
 			}
 		}
 		/* update */
-		if(CH->block_fnum != block_fnum)
+		if(CH->block_fnum != (UINT32) block_fnum)
 		{
 			UINT8 block  = block_fnum >> 10;
