+++ id_ca.cpp
@@ -480,7 +480,7 @@ void CAL_SetupGrFile (void)
     long headersize = tell(handle);
     lseek(handle, 0, SEEK_SET);
 
-    if(!param_ignorenumchunks && headersize / 3 != lengthof(grstarts) - numEpisodesMissing)
+    if(!param_ignorenumchunks && headersize / 3 != (long) (lengthof(grstarts) - numEpisodesMissing))
         Quit("Wolf4SDL was not compiled for these data files:\n"
             "%s contains a wrong number of offsets (%i instead of %i)!\n\n"
             "Please check whether you are using the right executable!\n"
